Restructure project layout
This fixes the problem where `sh -c flake8` was not working due to the strict way the entrypoint was designed.

Now the entrypoint is renamed to `insider`, and it acts as an entrypoint only to the insider scripts.

Renamed inside to insider and outside to outsider, to avoid confusing syntaxes.