Sort variables and remove $COMPOSE_FILE
The setting for COMPOSE_FILE introduced in dd77c326684943919abc30ea8131139f298488fc actually broke it, because it specified *all* the files that should be used, not *one of them in this order*. Oops, misunderstanding its meaning!

So now we just let the default COMPOSE_FILE behavior and require the user to be smarter than us by either setting the variable or adding a docker-compose.yml file.