Good separate example commands for test and devel
The previous command did not work in devel because it was trying to mount the `/project` folder from the host (where the docker socket was found), which was creating garbage there and simply not working. Besides, one couldn't copy-paste it into Fish.

Now we unrequire the `/project` path and make scripts more relocatable. Also we don't supply a default `COMPOSE_PROJECT_NAME`, so we let docker compose's default logic to do its magic when relocating the WORKDIR.

Since all of this does not make excessive sense in a CI environment, we're simply splitting the example commands here too.