Exclude module __init__.py files
All of them were getting reported as 0% coverage, which is basically impossible.

The problem is that they are loaded before the script is ran. Since they contain mostly no important code, we're excluding them here from the coverage tool.

This is not exactly a good solution, but it's a temporary improvement that fakes a pre-faked report to make it possible to reach 100% coverage.

See https://stackoverflow.com/a/16513048/1468388 for details.