Remove QA volume hacks
According to #8, I'm removing the QA volume transfer among containers.

This change requires that https://github.com/Tecnativa/doodba/pull/192 is properly working, merged and built.

It is a major milestone because this will allow QA testing in swarm/k8s environments, or anywhere where sharing volumes becomes a problem.

Also it will make all insider tests faster, and less garbage will be created in the underlying docker engine.

This is because now all dependencies needed by insider scripts are bundled in doodba itself, and it also has the new `/qa/insider` script, which lets you pass in a script as its 1st argument, and it will return the raw zip contents of `/qa/artifacts`. The data transfer is then done via stdin/stdout, which removes the need for an additional volume.

Includes the removal of `$SHARED_NETWORK`, which is not needed anymore
because the required external networks are extracted from the
docker compose file instead.