Remove QA volume hacks, networks-autocreate, add CI, and more (#9)
# Remove QA volume hacks

According to #8, I'm removing the QA volume transfer among containers.

This change requires that https://github.com/Tecnativa/doodba/pull/192 is properly working, merged and built.

It is a major milestone because this will allow QA testing in swarm/k8s environments, or anywhere where sharing volumes becomes a problem.

Also it will make all insider tests faster, and less garbage will be created in the underlying docker engine.

This is because now all dependencies needed by insider scripts are bundled in doodba itself, and it also has the new `/qa/insider` script, which lets you pass in a script as its 1st argument, and it will return the raw zip contents of `/qa/artifacts`. The data transfer is then done via stdin/stdout, which removes the need for an additional volume.

# networks-autocreate

The usage of `$SHARED_NETWORK`, is removed. The brand new `networks-autocreate` script ensures required external networks are present, and they are now extracted from the docker compose file directly instead.

# Improve .gitlab-ci.yml example

# Add CI to the project itself

# Do not fail if database exists previously in Odoo 8.0

This is needed because recent versions of the official postgres docker image let the developer choose the name of the initial database to be created, so it may exist previously. In fact, it defaults to the chosen postgres user, so possibly the "odoo" database exists before, and if that one is being used, then this command would fail with:

    createdb: database creation failed: ERROR:  database "odoo" already exists