Update .gitlab-ci.yml example to make it includable
This cannot be used directly in gitlab.com with the free runners, because it needs a persistent docker daemon across jobs. You need to use a custom runner instead, properly configured for these tasks.

If you use the proper runner, however, with these changes, you can directly include the file and make it work with just a few overrides.

It provides some sane defaults for most needed tasks. It will grow, however.