[CI] Fix build adapting to latest pylint-odoo behavior
Since https://github.com/OCA/pylint-odoo/pull/208, the version matcher in pylint-odoo only will validate a proper version number with the specified odoo version prefix.

The tests were hacking that bug, but that's no longer the case.

Now, a good addon structure is present for each version, and it is enabled in a specific `addons.yaml` file, which BTW serves us to check that the `ONLY` filtering works properly.