Use a image tag per pipeline
In an environment where images are shared, and where all pipelines were building the same image tag, several parallel pipelines could be conflicting each other and using another pipeline's image.

With this change, a mocked odoo image is built for each pipeline. In the jobs where the real odoo image name is needed (basically, where we push it to a registry), the mock is removed manually and images are rebuilt with cache (which is the same effect as re-tagging).