Avoid unexpected rebuilds in CI
I was expecting that Docker used the cache just by checking among its present images, but actually it's always doing the full rebuild. This is not desired when the job is tagged as "with cache", or when pushing the current image to the registry.

Now, the cache-enabled jobs should actually use cache, without impacting the builds without cache, and avoiding races between parallel jobs as much as possible.