Fix code to reuse pipeline image
The code from 41ce3a5e8b906cbbb902578f56194e9b789a0cb5 wasn't getting properly the fake image name. For instance, instead of `myproject-odoo:10.0` it was getting `myproject-odoo:latest`.

Besides, the code is simplified to avoid a multiline script that would be harder to debug in Gitlab CI logs.

Now, the pipeline image is detected with the same system as the real image. If there's no real image, it will fail. If re-tagging fails, the process will not fail.