Disable PYTHONOPTIMIZE by default
If a test has an `assert` statement, it would fail. Assertions are not meant for demos or production environments, but they can be useful for tests, for example when developing a base addon for other addons, and checking that its assertions will warn downstream development.

I'm disabling `$PYTHONOPTIMIZE` by default. Add it by setting the env var at will.