Fix false positive when wrong $ADDON_CATEGORIES is used
`coverage` is a special script because it doesn't start with `set -e` (on purpose; if it fails, it should continue).

Thus, to check that the addons listing process worked fine, we need to explicitly exit in case the return code isn't 0.

Also, since the same exit code (2) was being used for addons not found and for unrecognized arguments, in https://github.com/Tecnativa/doodba/pull/274 I modified addons not found to exit with code 4, and here I adapt all scripts to support that.

I also add tests for v13.

TT21246